Diving the code into modules
