+++ public/index.html
@@ -42,6 +42,6 @@
       </div>
     </main>
 
-    <script src="../src/main.js"></script>
+    <script type="module" src="./bundle.js"></script>
   </body>
 </html>
