+++ src/components/trip-events-list.js
@@ -1,0 +1,13 @@
+const createTripEventsList = () => {
+  return (
+    `<ul class="trip-days">
+      <li class="trip-days__item  day">
+        <div class="day__info"></div>
+        <ul class="trip-events__list">
+        </ul>
+      </li>
+    </ul>`
+  );
+};
+
+export {createTripEventsList};
